+++ admin/src/scenes/centers/youngs/index.js
@@ -14,6 +14,7 @@ import PencilAlt from "../../../assets/icons/PencilAlt";
 import ShieldCheck from "../../../assets/icons/ShieldCheck";
 import Template from "../../../assets/icons/Template.js";
 import SelectAction from "../../../components/SelectAction";
+import WithTooltip from "../../../components/WithTooltip";
 import api from "../../../services/api";
 import {
   canSearchInElasticSearch,
@@ -398,12 +399,14 @@ export default function CenterYoungIndex() {
         <div className="flex items-center justify-between">
           <div className="font-bold text-2xl mb-4">Volontaires</div>
           <div className="flex items-center gap-2">
-            <button
-              disabled={loading}
-              onClick={() => viewAttestation()}
-              className="flex justify-between items-center gap-3 px-3 py-2 rounded-lg cursor-pointer disabled:opacity-50 disabled:cursor-wait bg-blue-600 text-white font-medium text-sm">
-              Exporter les attestations
-            </button>
+            <WithTooltip tooltipText="Suite au remaniement ministériel du 4 juillet 2022, les nouvelles attestations ne sont pas encore disponibles. Elles le seront très prochainement">
+              <button
+                disabled={true}
+                // onClick={() => viewAttestation()}
+                className="flex justify-between items-center gap-3 px-3 py-2 rounded-lg cursor-pointer disabled:opacity-50 bg-blue-600 text-white font-medium text-sm">
+                Exporter les attestations
+              </button>
+            </WithTooltip>
             <SelectAction
               title="Exporter les volontaires"
               alignItems="right"