+++ api/src/index.js
@@ -42,9 +42,9 @@ function handleError(err, req, res, next) {
 
 const origin = [APP_URL, ADMIN_URL, SUPPORT_URL, KNOWLEDGEBASE_URL, "https://inscription.snu.gouv.fr"];
 app.use(cors({ credentials: true, origin }));
-app.use(bodyParser.json());
-app.use(bodyParser.text({ type: "application/x-ndjson" }));
-app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.json({ limit: "50mb" }));
+app.use(bodyParser.text({ limit: "50mb", type: "application/x-ndjson" }));
+app.use(bodyParser.urlencoded({ limit: "50mb", extended: true }));
 
 require("./crons");
 app.use(cookieParser());