+++ app/src/scenes/missions/viewDesktop.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { toastr } from "react-redux-toastr";
 import { useHistory, useParams } from "react-router-dom";
 import DoubleDayTile from "../../components/DoubleDayTile";
@@ -7,7 +7,17 @@ import Loader from "../../components/Loader";
 import WithTooltip from "../../components/WithTooltip";
 import api from "../../services/api";
 import plausibleEvent from "../../services/plausible";
-import { formatStringDateTimezoneUTC, htmlCleaner, translate, translateApplication, copyToClipboard, APPLICATION_STATUS, SENDINBLUE_TEMPLATES } from "../../utils";
+import {
+  formatStringDateTimezoneUTC,
+  htmlCleaner,
+  translate,
+  translateApplication,
+  copyToClipboard,
+  APPLICATION_STATUS,
+  SENDINBLUE_TEMPLATES,
+  translateAddFilePhase2,
+  COHESION_STAY_END,
+} from "../../utils";
 import DocumentsPM from "../militaryPreparation/components/DocumentsPM";
 import ApplyDoneModal from "./components/ApplyDoneModal";
 import ApplyModal from "./components/ApplyModal";
@@ -22,6 +32,12 @@ import rubberStampNotValided from "../../assets/rubberStampNotValided.svg";
 import { HiChevronDown, HiOutlineEye, HiOutlineMail, HiOutlineDownload } from "react-icons/hi";
 import ModalConfirm from "../../components/modals/ModalConfirm";
 import downloadPDF from "../../utils/download-pdf";
+import { AiOutlineClockCircle } from "react-icons/ai";
+import ModalPJ from "./components/ModalPJ";
+import { BsChevronDown } from "react-icons/bs";
+import { HiPlus } from "react-icons/hi";
+import FileCard from "./../../scenes/militaryPreparation/components/FileCard";
+
 
 export default function viewDesktop() {
   const [mission, setMission] = useState();
@@ -32,6 +48,8 @@ export default function viewDesktop() {
   const [loading, setLoading] = useState(false);
   const [contract, setContract] = useState(null);
   const [openContractDropdown, setOpenContractDropdown] = useState();
+  const [modalDocument, setModalDocument] = useState({ isOpen: false });
+  const [openAttachments, setOpenAttachments] = useState(false);
 
   const history = useHistory();
 
@@ -44,6 +62,8 @@ export default function viewDesktop() {
     const { data } = await api.get(`/mission/${id}`);
     return setMission(data);
   };
+
+  const optionsType = ["contractAvenantFiles", "justificatifsFiles", "feedBackExperienceFiles", "othersFiles"];
 
   useEffect(() => {
     const getContract = async () => {
@@ -189,6 +209,7 @@ export default function viewDesktop() {
               disabledIncomplete={disabledIncomplete}
               disabledPmRefused={disabledPmRefused}
               scrollToBottom={scrollToBottom}
+              young={young}
             />
           )}
         </div>
@@ -295,7 +316,6 @@ export default function viewDesktop() {
           <Detail title="Lieu" content={[mission.address, mission.zip, mission.city, mission.department]} />
         </div>
       </div>
-
       {mission.isMilitaryPreparation === "true" ? (
         <>
           <hr className="text-gray-100" />
@@ -304,6 +324,88 @@ export default function viewDesktop() {
           </div>
         </>
       ) : null}
+      {mission?.application?.status === "VALIDATED" && (
+        <>
+          <hr className="text-gray-100" />
+          <div className="mx-8 mt-8">
+            <div className="flex justify-between">
+              <div className="text-lg leading-6 font-semibold">Pièces jointes</div>
+              <div className="flex space-x-4 items-center">
+                {optionsType.reduce((nmb, option) => nmb + mission.application[option].length, 0) !== 0 && (
+                  <div
+                    className="group flex items-center rounded-lg text-blue-600 text-center text-sm py-2 px-10 border-blue-600 border-[1px] hover:bg-blue-600 hover:text-white transition duration-100 ease-in-out"
+                    onClick={() => setOpenAttachments(!openAttachments)}>
+                    Voir mes pièces jointes
+                    <BsChevronDown className={`ml-3 text-blue-600 group-hover:text-white h-5 w-5 ${openAttachments ? "rotate-180" : ""}`} />
+                  </div>
+                )}
+                <div
+                  className="text-white bg-blue-600  rounded-full p-2 "
+                  onClick={() => {
+                    setModalDocument({
+                      isOpen: true,
+                      stepOne: true,
+                    });
+                  }}>
+                  <HiPlus />
+                </div>
+              </div>
+            </div>
+            {openAttachments && (
+              <div className="flex flex-row overflow-x-auto gap-4 my-4 w-full ">
+                {optionsType.map(
+                  (option, index) =>
+                    mission.application[option].length > 0 && (
+                      <FileCard
+                        key={index}
+                        name={translateAddFilePhase2(option)[3].toUpperCase() + translateAddFilePhase2(option).slice(4)}
+                        icon="reglement"
+                        filled={mission.application[option].length}
+                        color="text-blue-600 bg-white"
+                        status="Modifier"
+                        onClick={() =>
+                          setModalDocument({
+                            isOpen: true,
+                            name: option,
+                            stepOne: false,
+                          })
+                        }
+                      />
+                    ),
+                )}
+              </div>
+            )}
+            <ModalPJ
+              isOpen={modalDocument?.isOpen}
+              name={modalDocument?.name}
+              young={young}
+              application={mission.application}
+              optionsType={optionsType}
+              onCancel={async () => {
+                setModalDocument({ isOpen: false });
+                await getMission();
+              }}
+              onSend={async (type, multipleDocument) => {
+                try {
+                  const responseNotification = await api.post(`/application/${mission.application._id}/notify/${SENDINBLUE_TEMPLATES.ATTACHEMENT_PHASE_2_APPLICATION}`, {
+                    type,
+                    multipleDocument,
+                  });
+                  if (!responseNotification?.ok) return toastr.error(translate(responseNotification?.code), "Une erreur s'est produite avec le service de notification.");
+                  toastr.success("L'email a bien été envoyé");
+                } catch (e) {
+                  toastr.error("Une erreur est survenue lors de l'envoi du mail", e.message);
+                }
+              }}
+              onSave={async () => {
+                setModalDocument({ isOpen: false });
+                await getMission();
+              }}
+              typeChose={modalDocument?.stepOne}
+            />
+          </div>
+        </>
+      )}
       {modal === "APPLY" && (
         <ApplyModal
           value={mission}
@@ -320,7 +422,20 @@ export default function viewDesktop() {
   );
 }
 
-const ApplyButton = ({ placesLeft, setModal, disabledAge, disabledIncomplete, disabledPmRefused, scrollToBottom }) => {
+const ApplyButton = ({ placesLeft, setModal, disabledAge, disabledIncomplete, disabledPmRefused, scrollToBottom, young }) => {
+  const now = new Date();
+  if (now < COHESION_STAY_END[young.cohort])
+    return (
+      <div className="flex flex-col items-center gap-2">
+        <WithTooltip tooltipText="Pour candidater, vous devez avoir terminé votre séjour de cohésion">
+          <button disabled className="px-12 py-2 rounded-lg text-white bg-blue-600 disabled:bg-blue-600/60 text-sm cursor-pointer">
+            Candidater
+          </button>
+        </WithTooltip>
+        <div className="text-xs leading-none font-normal text-gray-500">{placesLeft} places disponibles</div>
+      </div>
+    );
+
   if (disabledAge)
     return (
       <div className="flex flex-col items-center gap-2">